G4.M0: walking skeleton foundation — server+shared core, native cosine, cross-OS CI (#1)
Foundation spine for G4 Implement (THE-176).

shared: error taxonomy (codes + retryable), scope catalog (wildcard + HITL floor), config Zod schema, tool-result envelope.
server: adapter-agnostic Database interface + transactional sha256 migration runner (idempotent, drift = conflict); glob folder ACL (last-match-wins); canonical 16-byte argsHash (elicit + idempotency); audit event_log; ToolRegistry dispatch (validate -> auth -> scope/ACL -> HITL elicit -> execute -> byte governor -> audit); server_health admin tool. 28 vitest tests run under Node via node:sqlite.
native: napi-rs cosine_similarity (Vec<f64>) + build.rs calling napi_build::setup() for macOS linking.
CI: ci-server (biome lint + build-test matrix on ubuntu/macos/windows), ci-native (linux + macOS x64/arm64 + windows), ci-plugin (guarded until M4). Adds biome.json, strict tsconfig, committed bun.lock, .gitattributes (LF normalization).

Deferred to follow-on work (mechanical wiring against locked interfaces): concrete DB adapters (better-sqlite3 / bun:sqlite), MCP STDIO + Hono Streamable HTTP transports, jwt/ACL enforcement at the transport edge, HITL elicit round-trip, server_health STDIO smoke.

+++ packages/native/src/lib.rs
@@ -2,19 +2,21 @@
 
 use napi_derive::napi;
 
-/// Cosine similarity between two equal-length f32 vectors.
+/// Cosine similarity between two equal-length vectors.
 ///
-/// Placeholder implementation. The full vector primitives land in G2.3 (Storage
-/// schema) and G2.5 (Release engineering). This exists only to verify the
-/// napi-rs build pipeline works end-to-end.
+/// Placeholder implementation. Inputs are plain JS number arrays (f64); the
+/// zero-copy `Float32Array` path lands with the real vector primitives in G2.3
+/// (Storage schema) and G2.5 (Release engineering). napi-rs does not implement
+/// `FromNapiValue` for `Vec<f32>` (JS numbers are f64), so the binding takes
+/// `Vec<f64>`. This exists only to verify the napi-rs build pipeline end-to-end.
 #[napi]
-pub fn cosine_similarity(a: Vec<f32>, b: Vec<f32>) -> f32 {
+pub fn cosine_similarity(a: Vec<f64>, b: Vec<f64>) -> f64 {
     if a.len() != b.len() || a.is_empty() {
         return 0.0;
     }
-    let mut dot = 0.0_f32;
-    let mut norm_a = 0.0_f32;
-    let mut norm_b = 0.0_f32;
+    let mut dot = 0.0_f64;
+    let mut norm_a = 0.0_f64;
+    let mut norm_b = 0.0_f64;
     for i in 0..a.len() {
         dot += a[i] * b[i];
         norm_a += a[i] * a[i];
